Renaming package 'kenisis' to 'kinesis'
