+++ cmd/tndx-kinesis-error-redrive/main.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/rmrfslashbin/tndx/pkg/kenisis"
+	"github.com/rmrfslashbin/tndx/pkg/kinesis"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -35,7 +35,7 @@ var (
 	}
 	jsonFile string
 	log      *logrus.Logger
-	k        *kenisis.Config
+	k        *kinesis.Config
 )
 
 func init() {
@@ -45,10 +45,10 @@ func init() {
 		FullTimestamp: true,
 	})
 
-	k = kenisis.NewFirehose(
-		kenisis.SetRegion(region),
-		kenisis.SetLogger(log),
-		kenisis.SetDeliveryStream(deliveryStream),
+	k = kinesis.NewFirehose(
+		kinesis.SetRegion(region),
+		kinesis.SetLogger(log),
+		kinesis.SetDeliveryStream(deliveryStream),
 	)
 
 	RootCmd.Flags().StringVarP(&jsonFile, "jsonfile", "j", "", "path to json file")
