+++ cmd/tndx-kinesis-push/main.go
@@ -7,7 +7,7 @@ import (
 	"path"
 
 	"github.com/dghubble/go-twitter/twitter"
-	"github.com/rmrfslashbin/tndx/pkg/kenisis"
+	"github.com/rmrfslashbin/tndx/pkg/kinesis"
 	"github.com/rmrfslashbin/tndx/pkg/service"
 	"github.com/sirupsen/logrus"
 )
@@ -25,10 +25,10 @@ func main() {
 		FullTimestamp: true,
 	})
 
-	k := kenisis.NewFirehose(
-		kenisis.SetRegion(region),
-		kenisis.SetLogger(log),
-		kenisis.SetDeliveryStream(deliveryStream),
+	k := kinesis.NewFirehose(
+		kinesis.SetRegion(region),
+		kinesis.SetLogger(log),
+		kinesis.SetDeliveryStream(deliveryStream),
 	)
 
 	files, err := ioutil.ReadDir(jsonPath)
